move test to the right file
